Add files via upload
This program can tell a number between 1 and 100,000 that has the largest number of divisors and also it can tell the number of divisor this number(s) has. Since there can be numbers that have the same maximum number of divisors they will all be printed since this program use an array.