Add files via upload
This program animate colored disks.  New disks are continually added in the screen until the window is closed.  The positions of the disks are selected at random, and the color is randomly selected from red, green, blue, yellow, magenta, or cyan.  A black outline is drawn around each disk to emphasize the shape.